Initial release
Arduino code to drive common-cathode RGB LED's, via MOSFET drivers. Code uses a smooth sinusoidal PWM drive for a good-looking colour transition, based on the idea presented here: https://www.instructables.com/id/How-to-Make-Proper-Rainbow-and-Random-Colors-With-/

Incorporates a debounced interrupt-driven switch to toggle between RGB cycle and constant white (all LED's on).

The wiring to the Arduino UNO and circuit details for the MOSFET drivers are also provided.